**Happy New Year!** 🎉 

This PR adds a new `-filter-cf-errors` flag to httprobe which causes it to read the first 512 bytes from listening servers in order to determine if they are a "wildcard" response from Cloudflare like the ones shown here:

![Cloudflare error pages](https://i.imgur.com/QPk7vSt.png)

If a Cloudflare signature string is found, the function returns false to treat it as not listening. The functionality is implemented in a generic fashion so that it can easily be extended to filter out other common false-positive responses.

The PR also implements a small performance improvement by performing a DNS lookup on incoming domains and ignore the unresolvable ones to avoid filling up the job channels with dead domains. I timed the execution with a list containing 1483 resolveable and unresolvable domains with the `large` port list and was able to shave off 7 minutes of execution time, which is not a lot, but also not insignificant:

```bash
cat hosts.txt | time httprobe -t 2000 -p large
httprobe -t 2000 -p large  10.73s user 10.15s system 0% cpu 47:13.08 total

cat hosts.txt | time new-httprobe -t 2000 -p large
new-httprobe -t 2000 -p large  12.39s user 11.03s system 0% cpu 40:10.77 total

cat hosts.txt | time new-httprobe -t 2000 -filter-cf-errors -p large
new-httprobe -t 2000 -filter-cf-errors -p large  13.13s user 10.86s system 0% cpu 40:00.91 total
```