@michenriksen 2021-01-05
@tomnomnom I came across the `httpx` project by Project Discovery and noticed that the tool has [support for filter strings](https://github.com/projectdiscovery/httpx#usage) with  the `-filter-string` and `-filter-regex` flags. Would you be interested in having a similar feature for httprobe? If so, I will gladly refactor this PR to implement more generic filtering instead of the Cloudflare-specific filtering. :)